requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.6
pillow~=11.2.1
tqdm~=4.67.1
